Added 64-bit support and some fixes

+++ loader/Main/Utils.cs
@@ -6,66 +6,38 @@ namespace PenguLoader.Main
 {
     static class Utils
     {
-        public static void OpenFolder(string path)
-        {
-            Process.Start(new ProcessStartInfo()
-            {
-                FileName = "explorer.exe",
-                Arguments = $"\"{path}\"",
-                UseShellExecute = true
-            });
-        }
+        public static void OpenFolder(string path) => Process.Start("explorer.exe", $"\"{path}\"");
 
-        public static void OpenLink(string url)
-        {
-            Process.Start(url);
-        }
-
-        public static void RemoveAdminPerm(string path)
-        {
-            try
-            {
-                Process.Start(new ProcessStartInfo
-                {
-                    FileName = "cmd.exe",
-                    Arguments = $"/c echo Y| cacls \"{path}\" /grant \"{Environment.UserName}\":f",
-                    UseShellExecute = true,
-                    CreateNoWindow = true,
-                    Verb = "runas",
-                    WindowStyle = ProcessWindowStyle.Hidden
-                });
-            }
-            catch { }
-        }
+        public static void OpenLink(string url) => Process.Start(new ProcessStartInfo(url) { UseShellExecute = true });
 
         public static bool IsFileInUse(string path)
         {
+            if (!File.Exists(path)) return false;
+
             try
             {
-                if (!File.Exists(path))
-                    return false;
-
-                using (new FileStream(path, FileMode.Open, FileAccess.ReadWrite, FileShare.None))
-                {
-                    return false;
-                }
+                using (new FileStream(path, FileMode.Open, FileAccess.ReadWrite, FileShare.None)) { }
             }
             catch
             {
                 return true;
             }
+
+            return false;
         }
 
         public static void DeletePath(string path, bool isDir = false)
         {
             try
             {
-                if (isDir && Directory.Exists(path))
-                    Directory.Delete(path, true);
-                else if (File.Exists(path))
-                    File.Delete(path);
+                if (isDir) Directory.Delete(path, true);
+                else File.Delete(path);
             }
             catch { }
         }
+
+        public static void EnsureDirectoryExists(string path) => Directory.CreateDirectory(path);
+
+        public static void EnsureFileExists(string path) => File.Create(path).Close();
     }
 }
